requirements.txt
pandas
requests
beautifulsoup4
numpy
matplotlib
scipy
scikit-learn
python-chess
zstandard
black        # formateur de code
isort        # trier les imports alphabétiquement
mypy         # vérifier les types statiques
pre-commit   # orchestrateur qui lance tout avant git commit
